Extract assertCallableThrowsApiProblemException() to trait
